--- src/main/java/org/ohdsi/circe/cohortdefinition/builders/CriteriaBuilder.java
@@ -1,8 +1,0 @@
-package org.ohdsi.circe.cohortdefinition.builders;
-
-import org.ohdsi.circe.cohortdefinition.Criteria;
-
-public interface CriteriaBuilder<T extends Criteria> {
-
-    String getCriteriaSql(T criteria);
-}
